remote control: Add `/SESSION/CREATE` command
This allows creating a new empty session. This is useful for projects
that want to dynamically create the entire session using a custom module.

Without this, you can also create an empty session json file and load
that with `--load`, but then you would have to manage that file, and if
you accidentaly press control-s, it gets overwritten to contain any
dynamically created widgets.

With this commit, such a custom module can just create a new session on
startup and then populate it. This takes a bit of fiddling to get the
order of events correct, this commit also adds an example showing how to
do this.